+++ src/titiler/core/titiler/core/factory.py
@@ -174,11 +174,14 @@ class BaseTilerFactory(metaclass=abc.ABCMeta):
 
     def __post_init__(self):
         """Post Init: register route and configure specific options."""
+        # Register endpoints
         self.register_routes()
 
+        # Register Extensions
         for ext in self.extensions:
             ext.register(self)
 
+        # Update endpoints dependencies
         for scopes, dependencies in self.route_dependencies:
             self.add_route_dependencies(scopes=scopes, dependencies=dependencies)